procurement_agent/requirements.txt
google-adk
python-dotenv
tabulate
